Modify auto-experiment scripts to run aggregated test cases

+++ scripts/pokeEMU.sh
@@ -22,13 +22,13 @@ if ! [ -e $dir/single-m3-kvm/ ]; then
 	ssh yan@logan.cs.umn.edu $(echo "mv $remote_out $remote_dir/single-m3-qemu/")
 fi
 
-# Copy list of valid test cases for aggregating
-if ! [ -e aggreg_list ]; then
+# Generate the list of valid test cases for aggregating
+if ! [ -e aggreg_list ] && ! [ -e $dir/aggreg-m3-kvm/ ]; then
 	mkdir -p aggreg_list
 	for insn in $dir/single-m3-kvm/*; do
 		mkdir aggreg_list/$(basename $insn)
 		for file in $insn/*.diff; do
-			echo $remote_in/$(basename $file .diff)/testcase >> aggreg_list/$(basename $insn)/log
+			echo $remote_in/$(basename $insn)/$(basename $file .diff)/testcase >> aggreg_list/$(basename $insn)/log
 		done
 		var=$(tr "\n" "," < aggreg_list/$(basename $insn)/log)
 		echo ${var::-1} > aggreg_list/$(basename $insn)/log
@@ -36,4 +36,19 @@ if ! [ -e aggreg_list ]; then
 		rm aggreg_list/$(basename $insn)/log
 	done
 	scp -r aggreg_list/ yan@logan.cs.umn.edu:$remote_dir/aggreg_list/
+	rm -r aggreg_list/
 fi
+
+echo 'Run aggregate mode 3 experiment'
+if ! [ -e $dir/aggreg-m3-kvm/ ]; then
+	ssh yan@logan.cs.umn.edu $(echo "cd ~/Project/pokemu-oras/scripts; ./run-testcase.sh -aggreg -m 3 -in $remote_dir/aggreg_list/ -out $remote_out")
+	./run-testcase.sh -kvm -in $remote_out -out $out
+	mv $out $dir/aggreg-m3-kvm/
+	ssh yan@logan.cs.umn.edu $(echo "mv $remote_out $remote_dir/aggreg-m3-qemu/")
+fi
+
+echo 'generate effectiveness results'
+
+
+
+
